+++ lib/src/widgets/widgets.dart
@@ -1,3 +1,4 @@
 export 'adaptive_ui.dart';
 export 'gradient_widget.dart';
 export 'multi_tap_detector.dart';
+export 'typed_list_view.dart';